// src/app/models/ieventclasses.ts
export interface IEventClasses {

  _id: string,
  name: string;
  isValid: { type: Boolean; default: true };

  // BOYS
  boys6Class: string;
  boys6Fee: number;
  boys7Class: string;
  boys7Fee: number;
  boys8Class: string;
  boys8Fee: number;
  boys9Class: string;
  boys9Fee: number;
  boys10Class: string;
  boys10Fee: number;
  boys11Class: string;
  boys11Fee: number;
  boys12Class: string;
  boys12Fee: number;
  boys13Class: string;
  boys13Fee: number;
  boys14Class: string;
  boys14Fee: number;
  boys15Class: string;
  boys15Fee: number;
  boys16Class: string;
  boys16Fee: number;
  men17Class: string;
  men17Fee: number;
  men25Class: string;
  men25Fee: number;
  men30Class: string;
  men30Fee: number;
  men35Class: string;
  men35Fee: number;
  menJuniorClass: string;
  menJuniorFee: number;
  menUnder23Class: string;
  menUnder23Fee: number;
  menEliteClass: string;
  menEliteFee: number;
  // GIRLS
  girls7Class: string;
  girls7Fee: number;
  girls8Class: string;
  girls8Fee: number;
  girls9Class: string;
  girls9Fee: number;
  girls10Class: string;
  girls10Fee: number;
  girls11Class: string;
  girls11Fee: number;
  girls12Class: string;
  girls12Fee: number;
  girls13Class: string;
  girls13Fee: number;
  girls14Class: string;
  girls14Fee: number;
  girls15Class: string;
  girls15Fee: number;
  girls16Class: string;
  girls16Fee: number;
  women17Class: string;
  women17Fee: number;
  women25Class: string;
  women25Fee: number;
  womenJuniorClass: string;
  womenJuniorFee: number;
  womenUnder23Class: string;
  womenUnder23Fee: number;
  womenEliteClass: string;
  womenEliteFee: number;
  // CRUISER
  boys12CrClass: string;
  boys12CrFree: number;
  boys13CrClass: string;
  boys13CrFee: number;
  boys15CrClass: string;
  boys15CrFee: number;
  men17CrClass: string;
  men17CrFee: number;
  men25CrClass: string;
  men25CrFee: number;
  men30CrClass: string;
  men30CrFee: number;
  men35CrClass: string;
  men35CrFee: number;
  men40CrClass: string;
  men40CrFee: number;
  men50CrClass: string;
  men50CrFee: number;

  girls12CrClass: string;
  girls12CrFee: number;
  girls13CrClass: string;
  girls13CrFee: number;
  women17CrClass: string;
  women17CrFee: number;
  women30CrClass: string;
  women30CrFee: number;
  women40CrClass: string;
  women40CrFee: number;
}